+++ src/wininput.c
@@ -22,11 +22,17 @@ static int sysmenulen;
 //static uint kb_select_key = 0;
 static uint super_key = 0;
 static uint hyper_key = 0;
+// Compose support
+static uint compose_key = 0;
+static uint last_key_down = 0;
+static uint last_key_up = 0;
+
 static uint newwin_key = 0;
 static bool newwin_pending = false;
 static bool newwin_shifted = false;
 static bool newwin_home = false;
 static int newwin_monix = 0, newwin_moniy = 0;
+
 static int transparency_pending = 0;
 static bool selection_pending = false;
 bool kb_input = false;
@@ -1438,6 +1444,14 @@ hyper_down(uint key, mod_keys mods)
 }
 
 static void
+compose_down(uint key, mod_keys mods)
+{
+  compose_key = key;
+  last_key_down = key;
+  (void)mods;
+}
+
+static void
 kb_select(uint key, mod_keys mods)
 {
   (void)mods;
@@ -1822,6 +1836,7 @@ static struct function_def cmd_defs[] = {
 
   {"super", {.fct_key = super_down}, 0},
   {"hyper", {.fct_key = hyper_down}, 0},
+  {"compose", {.fct_key = compose_down}, 0},
   {"kb-select", {.fct_key = kb_select}, mflags_kb_select},
   {"no-scroll", {.fct_key = no_scroll}, mflags_no_scroll},
   {"toggle-no-scroll", {.fct_key = toggle_no_scroll}, mflags_no_scroll},
@@ -1876,8 +1891,6 @@ typedef enum {
   COMP_PENDING = 1, COMP_ACTIVE = 2
 } comp_state_t;
 static comp_state_t comp_state = COMP_NONE;
-static uint last_key_down = 0;
-static uint last_key_up = 0;
 
 static struct {
   wchar kc[4];
@@ -2279,6 +2292,14 @@ pick_key_function(wstring key_commands, char * tag, int n, uint key, mod_keys mo
 
       return ret;
     }
+#ifdef support_compose_by_capslock_via_config
+#warning this causes endless event looping, to be worked out if desired...
+    else if (key == VK_CAPITAL && cfg.compose_key == MDK_CAPSLOCK) {
+      // nullify the keyboard state lock effect, see above,
+      // so we can/could support config ComposeKey=capslock
+      win_key_nullify(key);
+    }
+#endif
 
     n--;
     if (sepp) {
@@ -3746,7 +3767,7 @@ win_key_up(WPARAM wp, LPARAM lp)
 
   uint key = wp;
 #ifdef debug_virtual_key_codes
-  printf("  win_key_up %02X %s\n", key, vk_name(key));
+  printf("  win_key_up %02X (down %02X) %s\n", key, last_key_down, vk_name(key));
 #endif
 
   if (key == VK_CANCEL) {
@@ -3755,6 +3776,7 @@ win_key_up(WPARAM wp, LPARAM lp)
     // detected properly for use as a modifier; let's try to fix this
     super_key = 0;
     hyper_key = 0;
+    compose_key = 0;
   }
   else if (key == VK_SCROLL) {
     // heuristic compensation of race condition with auto-repeat
@@ -3782,6 +3804,13 @@ win_key_up(WPARAM wp, LPARAM lp)
         (cfg.compose_key == MDK_ALT && key == VK_MENU)
         || (cfg.compose_key == MDK_SUPER && key == super_key)
         || (cfg.compose_key == MDK_HYPER && key == hyper_key)
+        // support KeyFunctions=CapsLock:compose (or other key)
+        || key == compose_key
+#ifdef support_compose_by_capslock_via_config
+#warning needs support by nullifying capslock state (see above)
+        // support config ComposeKey=capslock
+        || (cfg.compose_key == MDK_CAPSLOCK && key == VK_CAPITAL)
+#endif
        )
     {
       if (comp_state >= 0) {